add e2e test instead
