+++ packages/x-data-grid/src/tests/rows.DataGrid.test.tsx
@@ -902,72 +902,6 @@ describe('<DataGrid /> - Rows', () => {
           });
         },
       );
-
-      // https://github.com/mui/mui-x/issues/14726
-      testSkipIf(isJSDOM).only(
-        'should not cause scroll jumping when the focused cell is outside of the viewport',
-        async () => {
-          const rows = [
-            { id: 0 },
-            { id: 1 },
-            { id: 2 },
-            { id: 3 },
-            { id: 4 },
-            { id: 5 },
-            { id: 6 },
-            { id: 7 },
-            { id: 8 },
-            { id: 9 },
-            { id: 10 },
-            { id: 11 },
-            { id: 12 },
-            { id: 13 },
-            { id: 14 },
-            { id: 15 },
-            { id: 16 },
-            { id: 17 },
-            { id: 18 },
-            { id: 19 },
-            { id: 20 },
-          ];
-          const { user } = render(
-            <div style={{ height: 300, width: 400 }}>
-              <DataGrid rows={rows} columns={[{ field: 'id' }]} getRowHeight={() => 'auto'} />
-            </div>,
-          );
-
-          await user.click(getCell(2, 0));
-
-          const virtualScroller = grid('virtualScroller')!;
-          const scrollPositions: number[] = [];
-          virtualScroller.addEventListener('scroll', () => {
-            scrollPositions.push(virtualScroller.scrollTop);
-          });
-
-          virtualScroller.scrollBy({ top: 150, left: 0, behavior: 'smooth' });
-          virtualScroller.dispatchEvent(new Event('scroll'));
-
-          await waitFor(() => {
-            expect(virtualScroller.scrollTop).to.equal(150);
-          });
-
-          const hasScrollJump = scrollPositions.some((scrollTop, index) => {
-            if (index === 0) {
-              return false;
-            }
-            // When scrolling down, the scrollTop should be always decreasing
-            const prevScrollTop = scrollPositions[index - 1];
-            return scrollTop < prevScrollTop;
-          });
-
-          console.log('scrollPositions', scrollPositions);
-
-          expect(hasScrollJump).to.equal(
-            false,
-            `Scroll jumped, scrollPositions: ${scrollPositions}`,
-          );
-        },
-      );
     });
   });
